requirements.txt
flask
numpy
pillow
tensorflow
joblib
scikit-learn
